python-ecosystem/rag-pipeline/requirements.txt
# Core dependencies
python-dotenv>=1.0.0,<2.0.0

# OpenAI client (for OpenRouter compatibility)
openai>=1.12.0,<2.0.0
tiktoken>=0.5.0,<1.0.0

# LlamaIndex core
llama-index-core==0.13.0

# LlamaIndex extensions
llama-index-embeddings-openai>=0.3.0
llama-index-vector-stores-qdrant>=0.5.0
llama-index-llms-openai>=0.3.0

# Qdrant for vector storage
qdrant-client>=1.7.0,<2.0.0

# Web framework
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# File handling
aiofiles>=23.2.0,<24.0.0

# Testing
pytest>=8.0.0,<9.0.0

# HTTP client for OpenRouter
httpx>=0.27.0,<1.0.0

